refactor: use utility
