+++ lib/v2/taptap/review.js
@@ -1,5 +1,6 @@
 const got = require('@/utils/got');
 const { parseDate } = require('@/utils/parse-date');
+const { rootUrl, X_UA } = require('./utils');
 
 const sortMap = {
     default: {
@@ -20,8 +21,6 @@ const sortMap = {
     },
 };
 
-const rootUrl = 'https://www.taptap.com';
-
 module.exports = async (ctx) => {
     const id = ctx.params.id;
     const order = ctx.params.order ?? 'default';
@@ -33,7 +32,7 @@ module.exports = async (ctx) => {
         sortParam = `sort=${order}`;
         url += `&${sortParam}`;
     }
-    url += `&X-UA=V=1%26PN=WebApp%26VN=0.1.0%26LANG=${lang}%26PLT=PC`;
+    url += `&${X_UA(lang)}`;
 
     const reviews_list_response = await got(url);
     const reviews_list = reviews_list_response.data;
