+++ lib/v2/taptap/topic.js
@@ -1,10 +1,7 @@
 const got = require('@/utils/got');
-const cheerio = require('cheerio');
 const { parseDate } = require('@/utils/parse-date');
-const path = require('path');
-const { art } = require('@/utils/render');
+const { rootUrl, X_UA, appId2GroupId, textPost, videoPost } = require('./utils');
 
-const rootUrl = 'https://www.taptap.com';
 const typeMap = {
     feed: {
         zh_CN: '最新',
@@ -24,38 +21,6 @@ const typeMap = {
     },
 };
 
-const appId2GroupId = async (appId, lang = 'zh_CN') => {
-    const res = await got(`${rootUrl}/webapiv2/group/v1/detail?app_id=${appId}&X-UA=V=1%26PN=WebApp%26VN=0.1.0%26LANG=${lang}%26PLT=PC`, {
-        headers: {
-            Referer: `${rootUrl}/app/${appId}`,
-        },
-    });
-    return res.data.data.group.id;
-};
-
-const textPost = async (appId, topicId, lang = 'zh_CN') => {
-    const res = await got(`${rootUrl}/webapiv2/topic/v1/detail?id=${topicId}&X-UA=V=1%26PN=WebApp%26VN=0.1.0%26LANG=${lang}%26PLT=PC`, {
-        headers: {
-            Referer: `${rootUrl}/app/${appId}`,
-        },
-    });
-    const $ = cheerio.load(res.data.data.first_post.contents.text);
-    $('img').each((_, e) => {
-        e = $(e);
-        e.attr('src', e.attr('data-origin-url'));
-        e.attr('referrerpolicy', 'no-referrer');
-        e.removeAttr('data-origin-url');
-    });
-    return $.html();
-};
-
-const videoPost = (video) =>
-    art(path.join(__dirname, 'templates/videoPost.art'), {
-        intro: video?.intro?.text,
-        videoUrl: video?.url,
-        posterUrl: video?.raw_cover.url,
-    });
-
 module.exports = async (ctx) => {
     const lang = ctx.params.lang ?? 'zh_CN';
     const appId = ctx.params.id;
@@ -63,7 +28,7 @@ module.exports = async (ctx) => {
     const type = ctx.params.type ?? 'feed';
     const sort = ctx.params.sort ?? 'created';
 
-    const url = `${rootUrl}/webapiv2/feed/v6/by-group?group_id=${groupId}&type=${type}&sort=${sort}&X-UA=V=1%26PN=WebApp%26VN=0.1.0%26LANG=${lang}%26PLT=PC`;
+    const url = `${rootUrl}/webapiv2/feed/v6/by-group?group_id=${groupId}&type=${type}&sort=${sort}&${X_UA(lang)}`;
 
     const topics_list_response = await got(url);
     const topics_list = topics_list_response.data;
@@ -82,6 +47,7 @@ module.exports = async (ctx) => {
         try {
             description = await textPost(appId, topicId, lang);
         } catch (e) {
+            // fallback if 400 bad request
             description = list.moment.extended_entities?.topics?.[0].summary || list.moment.contents.raw_text || videoPost(list.moment.extended_entities?.videos?.[0]);
         }
